rpm: do not fail on non-existing /etc/init/serial.conf
This file is obsolete for a long time, so use it only if found in the
system (perhaps still useful in other distros).

+++ rpm_spec/core-vm.spec
@@ -100,7 +100,9 @@ adduser --create-home user
 make install-vm DESTDIR=$RPM_BUILD_ROOT
 
 %triggerin -- initscripts
-cp /usr/lib/qubes/serial.conf /etc/init/serial.conf
+if [ -e /etc/init/serial.conf ]; then
+	cp /usr/lib/qubes/serial.conf /etc/init/serial.conf
+fi
 
 %post
